[Benchmark] Latex-based reporting; Factorize benchmark.py entri point; Minor fixes
