+++ benchmark/.gitignore
@@ -1,0 +1,4 @@
+_benchmark
+_img
+svg-inkscape
+datasets
